fix stack oveflow when trying to get html-resource
When attempting to get html/html-resource for a large resource that is
not a web-page but, e.g. a large mp4 video file we end up crashing IRC
thread (the one where callback is running) with
java.lang.StackOverflowError exception. To deal with that we first
perform HEAD request and proceed further only for 'text/html' content
type.