add client-side ping timeout
To be able to detect situations when we were disconnected, we replaced
default stdout-handler with a custom one, that will update timestamp
stored in atom, in addition to printing incoming message to stdout.

In separate thread, we are checking if time that has passed since last
incoming message is greater than timeout and try to reconnect, if
needed.